\name{DMMD-package}
\alias{DMMD-package}
%\alias{DMMD}
\docType{package}
\title{
DNA Methylation Motif Discovery (DMMD)
}
\description{
This package creates DNA methylation motifs
}
\details{
\tabular{ll}{
Package: \tab DMMD\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-06-24\cr
License: \tab GPL-2\cr
}
DNA Methylation Motifs
}
\author{
Who wrote it

Maintainer: Inigo Reiriz <inigoreiriz@gmail.com> and Marcos J. Arauzo-Bravo <mararabra@yahoo.co.uk>
}
\keyword{DNA Methylation Motif Discovery, Next Generation Sequencing}